Added random pass string

+++ src/actions/pass.js
@@ -8,7 +8,7 @@ module.exports = () => async (ctx) => {
 
   config = JSON.parse(config)
 
-  if (!ctx.session.restricted || ctx.from.id !== Number(ctx.match[1])) {
+  if (!ctx.session.restricted || ctx.from.id !== Number(ctx.match[2])) {
     return ctx.answerCbQuery(config.captcha.notYouToast)
   }
   ctx.session.restricted = null
@@ -18,18 +18,28 @@ module.exports = () => async (ctx) => {
     ctx.session.timeoutToKick = null
   }
 
-  await ctx.restrictChatMember(
-    ctx.from.id,
-    {
-      can_send_messages: true,
-      can_send_media_messages: true,
-      can_send_other_messages: true,
-      can_add_web_page_previews: true,
-    }
-  )
+  if (ctx.session.pass === ctx.match[1]) {
+    ctx.session.pass = null
+    await ctx.restrictChatMember(
+      ctx.from.id,
+      {
+        can_send_messages: true,
+        can_send_media_messages: true,
+        can_send_other_messages: true,
+        can_add_web_page_previews: true,
+      }
+    )
 
-  ctx.answerCbQuery(config.captcha.successToast)
-  await ctx.deleteMessage()
+    ctx.answerCbQuery(config.captcha.successToast)
+    return ctx.deleteMessage()
+  }
 
-  return true
+  ctx.answerCbQuery(config.captcha.failuteToast)
+  await ctx.tg.kickChatMember(ctx.chat.id, ctx.from.id)
+
+  setTimeout(() => {
+    ctx.tg.unbanChatMember(ctx.chat.id, ctx.from.id)
+  }, 40000)
+
+  return ctx.deleteMessage()
 }
